requirements.txt

asgiref==3.7.2
asttokens==2.2.1
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
Django==4.2.2
elementpath==4.1.2
executing==1.2.0
filelock==3.12.2
icecream==2.1.3
idna==3.4
image==1.5.33
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
opencv-python==4.7.0.72
pandas==2.0.2
Pillow==9.5.0
Pygments==2.15.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sqlparse==0.4.4
sympy==1.12
threadpoolctl==3.1.0
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.65.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3